+++ configurations/conemsstate.h
@@ -21,7 +21,7 @@ class ConEMSState
 {
     Q_GADGET
     Q_PROPERTY(QJsonObject currentState READ currentState WRITE setCurrentState USER true)
-    Q_PROPERTY(uint64_t timestamp READ timestamp WRITE setTimestamp USER true)
+    Q_PROPERTY(long long timestamp READ timestamp WRITE setTimestamp USER true)
 
 public:
     ConEMSState();
@@ -31,15 +31,15 @@ public:
     QJsonObject currentState() const;
     void setCurrentState(const QJsonObject currentState);
 
-    uint64_t timestamp() const;
-    void setTimestamp(const uint64_t timestamp);
+    long long timestamp() const;
+    void setTimestamp(const long long timestamp);
 
     bool operator==(const ConEMSState &other) const;
     bool operator!=(const ConEMSState &other) const;
 
 private:
     QJsonObject m_currentState = QJsonObject() ;
-    uint64_t m_timestamp = 0;
+    long long m_timestamp = 0;
